Problem: conditional structure members are always present
Given something like:

```rust
struct Foo {
  #[cfg(unix)]
  bar: i32
}
```

we will always get a structure with `bar` field, regardless of
conditions.

Solution: pass conditional configuration for fields
and generate conditional #ifdef for them as well.

Current solution isn't absolutely perfect as it still inserts
semicolons after the conditions; but at least in the meanwhile this
seems to be an adequate solution.