ir: repr: support #[repr(align(...))] and #[repr(packed)] parsing
While #[repr(packed)] matches the rest of the currently-implemented
reprs, #[repr(align(...))] adds another level of nesting and so it's
necessary to handle arguments for nested #[repr(...)]s. However,
arguments are only permitted for #[repr(align(...))] and any further
nesting is ignored as before.

It should be noted that #[repr(packed(...))] is also valid in Rust, but
due to some complications with translating it to C, it is currently
unsupported.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>