ir: struct: support #[repr(align(...))] and #[repr(packed)]
As with #[must_use], this new support requires users to specify what
annotations they would like to apply when an alignment-related repr is
encountered. However, unlike #[must_use] all compilers appear to use the
same attributes to indicate layout settings so there is just a single
configuration for this new feature in a separate [layout] TOML table.

In order to facilitate the fact that #[repr(align(...))] takes an
argument which will likely need to be included in the final annotation,
the user must specify a function-like macro which takes the alignment
value as an argument.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>