{struct,union}: do not generate unsafe layouts
If the user has not specified a corresponding annotation for types that
carry #[repr(packed)] or #[repr(align(...))] we must not generate the
layout because it can cause interoperability issues (and possible memory
unsafety) between C and Rust code.

This unfortunately required some less-than-pretty plumbing changes so
that the load phase of the parser has access to the complete Config
object (luckily this does make the overall code look a bit cleaner
IMHO). The individual IR loaders only get their corresponding structs,
to avoid introducing confusing semantics.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>