ty: Simplify Box<T> to *T
This simplifies Box<T> to *T, which enables use cases not previously
supported.

Box<T> is guaranteed to have the same layout as *T as long as T is
sized. This is documented in the current Rust beta.

This does mean that the headers we're emitting are incorrect for unsized
types, but it still seems like a strict improvement over old behavior.

This can also be a recipe for memory leaks if not handled carefully, but
again, still seems better than previously.

Test cases are shamelessly copied and modified from the NonNull simplification.