Add rename rule for generated associated constant
In FFI Rust code it is often necessary to use constants instead of enums
for compatibility reasons. The most natural thing to do to somewhat
preserve grouping is to use associated constants. E.g.:

  ```rust
  #[repr(C)]
  struct Foo {}

  impl Foo {
      pub const FLAG1: u32 = 10;
      pub const FLAG2: u32 = 11;
      ...
  ```

For the above, the generated constants would be called Foo_FLAG1 &
Foo_FLAG2. Many Linux core C libraries adhere to a slightly different
casing, however, where the type prefix is fully upper cased: FOO_FLAG1.
Currently, it does not seem possible to generate such constants.

This change introduces a new config option to structured types that
enables this use case: the struct.rename_associated_constant option
expects a rename rule (as other rename related config options) and
applies it to the struct base name as used in conjunction with
associated constants to form the final name of the constant.

Signed-off-by: Daniel Müller <deso@posteo.net>