Add export config to expose all structs, enums or unions
Currently, there is no simple way to instruct cbindgen to simply export
all public structs, enums and/or unions, regardless of whether the items
themselves appear in public functions or not. In order to do this, each
item would need to be added manually to the `include` config item under
`[export]`.

This change adds coarse-grained config items under `[export]` to simply
export all structs, enums and/or unions. Items can further be excluded
with the `exclude` config item under the same category.

Signed-off-by: Carlos López <carlos.lopez@openchip.com>